Optionally disable index scan when replica identity is full
When replica identitiy is full, the logical replication apply workers
for subscription is capable of using non-unique index scans. However,
index scans has some overhead, especially when the data on the indexed
column(s) is duplicated.

We add a new option to subscription_parameter for both CREATE and ALTER
subscription commands such that the user can control the behavior.